+++ src/Twilio/Rest/Verify/V2/Service/WebhookList.php
@@ -56,6 +56,7 @@ class WebhookList extends ListResource {
             'EventTypes' => Serialize::map($eventTypes, function($e) { return $e; }),
             'WebhookUrl' => $webhookUrl,
             'Status' => $options['status'],
+            'Version' => $options['version'],
         ]);
 
         $payload = $this->version->create('POST', $this->uri, [], $data);